+++ collatex-pythonport/HISTORY.rst
@@ -3,7 +3,13 @@
 History
 -------
 
-2.0.0pre1 (2014-06-10)
+Development
+++++++++++++++++++
+
+* Removed max 6 witness limit in aligner, now n number of witnesses can be aligned. 
+
+2.0.0pre1 (2014-06-02)
 ++++++++++++++++++
 
 * First release on PyPI.
+* New collation algorithm, which does non progressive multiple witness alignment.